Text to sql Chainlit/requirements.txt
langchain
chainlit
openai
python-dotenv
